refactor: remove device.py and move all function to utils.py
