+++ faster_qwen3_tts/predictor_graph.py
@@ -17,7 +17,7 @@ import torch
 from transformers import StaticCache
 from transformers.masking_utils import create_causal_mask, create_sliding_window_causal_mask
 
-from .device import device_supports_cuda_graphs
+from .utils import device_supports_cuda_graphs
 from .sampling import sample_logits
