+++ faster_qwen3_tts/talker_graph.py
@@ -16,8 +16,7 @@ Strategy:
 import torch
 from transformers import StaticCache
 from transformers.masking_utils import create_causal_mask, create_sliding_window_causal_mask
-
-from .device import device_supports_cuda_graphs
+from .utils import device_supports_cuda_graphs
 
 
 class TalkerGraph:
